:sparkles: Add default configuration to limit auto-backup file size
Use reference copy for files exceeding auto-backup size limit
