+++ composeApp/src/desktopMain/kotlin/com/crosspaste/paste/item/FilesPasteItem.kt
@@ -9,11 +9,9 @@ import com.crosspaste.presist.DesktopOneFilePersist
 import com.crosspaste.presist.FileInfoTree
 import com.crosspaste.serializer.PathStringRealmListSerializer
 import com.crosspaste.serializer.StringRealmListSerializer
-import com.crosspaste.utils.DesktopFileUtils.createPasteRelativePath
 import com.crosspaste.utils.DesktopJsonUtils
 import io.realm.kotlin.MutableRealm
 import io.realm.kotlin.ext.realmListOf
-import io.realm.kotlin.ext.toRealmList
 import io.realm.kotlin.types.RealmList
 import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.annotations.Index
@@ -40,7 +38,7 @@ class FilesPasteItem : RealmObject, PasteItem, PasteFiles {
     var identifiers: RealmList<String> = realmListOf()
 
     @Serializable(with = PathStringRealmListSerializer::class)
-    var relativePathList: RealmList<String> = realmListOf()
+    override var relativePathList: RealmList<String> = realmListOf()
 
     var fileInfoTree: String = ""
 
@@ -66,10 +64,6 @@ class FilesPasteItem : RealmObject, PasteItem, PasteFiles {
         return AppFileType.FILE
     }
 
-    override fun getRelativePaths(): List<String> {
-        return relativePathList
-    }
-
     override fun getFilePaths(): List<Path> {
         val basePath = basePath?.toPath() ?: DesktopPathProvider.resolve(appFileType = getAppFileType())
         return relativePathList.map { relativePath ->
@@ -86,26 +80,6 @@ class FilesPasteItem : RealmObject, PasteItem, PasteFiles {
         return getFilePaths().flatMap { path ->
             val fileInfoTree = fileInfoTreeMap[path.name]!!
             fileInfoTree.getPasteFileList(path)
-        }
-    }
-
-    // use to adapt relative paths when relative is no storage in crossPaste
-    override fun adaptRelativePaths(
-        appInstanceId: String,
-        pasteId: Long,
-    ) {
-        val noStorageInCrossPaste = relativePathList.any { it.toPath().segments.size == 1 }
-        if (noStorageInCrossPaste) {
-            relativePathList =
-                relativePathList.map { relativePath ->
-                    val path = relativePath.toPath()
-                    val fileName = path.name
-                    createPasteRelativePath(
-                        appInstanceId = appInstanceId,
-                        pasteId = pasteId,
-                        fileName = fileName,
-                    )
-                }.toRealmList()
         }
     }
